requirements.txt
Django
gunicorn
strawberry-graphql
django-cors-headers
